Factor out REST API methods
